backend/requirements.txt
# AI Job Matcher — Backend Dependencies
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
psycopg2-binary>=2.9.9,<3.0.0
python-dotenv>=1.0.0,<2.0.0
google-generativeai>=0.8.0,<1.0.0
pydantic>=2.5.0,<3.0.0
python-multipart>=0.0.9,<1.0.0
openai>=1.0.0,<2.0.0
httpx>=0.27.0,<1.0.0
beautifulsoup4>=4.12.0,<5.0.0
python-docx>=1.1.0,<2.0.0
google-api-python-client>=2.100.0,<3.0.0
google-auth>=2.25.0,<3.0.0
apscheduler>=3.10.0,<4.0.0
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
pandas>=2.0.0,<3.0.0
openpyxl>=3.1.2,<4.0.0
